Issue/#6 メッセージ機能を実装しました

+++ db/migrate/20220209131522_create_relationships.rb
@@ -1,8 +1,8 @@
 class CreateRelationships < ActiveRecord::Migration[6.0]
   def change
     create_table :relationships do |t|
-      t.integer :following_id, null: false, foreign_key: true
-      t.integer :follower_id, null: false, foreign_key: true
+      t.bigint :following_id, null: false, foreign_key: true
+      t.bigint :follower_id, null: false, foreign_key: true
 
       t.timestamps
     end
